### 2018-09-28 #983 @whedon
[PRE REVIEW]: Pyoints: A Python package for point cloud, voxel and raster processing.
**Submitting author:** @laempy (<a href="http://orcid.org/0000-0002-8963-2762">Sebastian Lamprecht</a>)
**Repository:** <a href="https://github.com/laempy/pyoints" target ="_blank">https://github.com/laempy/pyoints</a>
**Version:** v0.2.0a2
