### 2018-09-28 #983 @lheagy
:wave: Hi @Chilipp, @kbevers, would you be willing to review this submission for JOSS?
