### 2018-09-30 #983 @lheagy
👋 Hi @Scivision would you be willing to review this submission for JOSS?
