+++ pyoints/storage/LasHandler.py
@@ -198,7 +198,7 @@ def writeLas(geoRecords, outfile):
     offset[:dim] = geoRecords.t.origin
 
     max_values = np.abs(records.extent().corners - offset[:dim]).max(0)
-    max_digits = 2**30  # long int
+    max_digits = 2**28 # long int
     scale[:dim] = max_values / max_digits
     scale[np.isclose(scale, 0)] = 1 / max_digits
 
