### 2018-10-09 laempy/pyoints#2 @kbevers
[JOSS-REVIEW] Installation instructions
In relation to https://github.com/openjournals/joss-reviews/issues/990

The installation instructions are only written for Ubuntu, although in the text it is somewhat hinted that Windows is also supported.

A clear description of which systems this package is intended for would be nice. Since this is a Python package it (theoretically) should work on at least Linux, OS X and Windows.

I recommend adding a conda environment file and basing the installation instructions on that instead. This will improve the chances of a successful installation immensely.

As a side note, I happen to work on a Mac and am having issues getting started with the installation. This package is not at fault, a mix of homebrew several GDAL and Python installations are complicating things a bit here. This could be avoided with a conda setup.
