### 2018-10-09 laempy/pyoints#3 @kbevers
[JOSS-REVIEW] Community guidelines missing
In relation to openjournals/joss-reviews#990

The community guidelines currently consists of these few words under the section "Contact":

	In case of questions, suggestions for improvement or software issues, please use
	the issue board or contact Sebastian Lamprecht, lamprecht@uni-trier.de.
	Contribution to the software is always welcome.

While this touches the bare minimum I would appreciate a separate section called "Community" or "Contributing", describing in a bit more words how to contribute to the project either with code or bug reports. GitHub has a nice feature for this in [contributing.md](https://blog.github.com/2012-09-17-contributing-guidelines/) files.
