### 2018-10-09 laempy/pyoints#4 @kbevers
[JOSS-REVIEW] Missing DOI in reference list
In relation to https://github.com/openjournals/joss-reviews/issues/990

The following reference is missing DOI's:

[Rusu, R. B., & Cousins, S. (2011). 3D is here: Point Cloud Library (PCL). In IEEE
International Conference on Robotics and Automation (ICRA). Shanghai, China.](http://www.scirp.org/(S(czeh2tfqyw2orz553k1w0r45))/reference/ReferencesPapers.aspx?ReferenceID=1671293)
