### 2018-10-09 laempy/pyoints#5 @kbevers
[JOSS-REVIEW] Citations in paper.md
In relation to https://github.com/openjournals/joss-reviews/issues/990

In text citation for various software packages are done in an incorrect manner. Instead of

> Libraries, like (Rusu & Cousins, 2011), (Zhou, Park, & Koltun, 2018) and (PDAL contributors, 2018) focus on 3D point cloud processing.

write

> Libraries, like PCL (Rusu & Cousins, 2011), Open3D (Zhou, Park, & Koltun, 2018) and PDAL (PDAL contributors, 2018) focus on 3D point cloud processing.

See [here](http://www.citethisforme.com/harvard-referencing#harvard-in-text-citations) for more examples of correct use of in-text citations.
