### 2018-10-12 c6b48ae
Classification flags added to LAS
