+++ pyoints/examples/file_examples/dump_example.py
@@ -34,7 +34,7 @@ Create GeoRecords from scratch.
 >>> print(geoRecords.shape)
 (1000,)
 >>> print(sorted(geoRecords.dtype.descr))
-[('classification', '<i8'), ('coords', '<f8', (3,)), ('intensity', '<i8'), ('values', '<f8')]
+[('classification', '<u8'), ('coords', '<f8', (3,)), ('intensity', '<u8'), ('keypoint', '|b1'), ('synthetic', '|b1'), ('values', '<f8'), ('withheld', '|b1')]
 
 Save as a DUMP-file.
 
@@ -49,6 +49,6 @@ Load the DUMP-file again and check the characteristics.
 >>> print(geoRecords.shape)
 (1000,)
 >>> print(sorted(geoRecords.dtype.descr))
-[('classification', '<i8'), ('coords', '<f8', (3,)), ('intensity', '<i8'), ('values', '<f8')]
+[('classification', '<u8'), ('coords', '<f8', (3,)), ('intensity', '<u8'), ('keypoint', '|b1'), ('synthetic', '|b1'), ('values', '<f8'), ('withheld', '|b1')]
 
 """
