### 2018-10-21 #990 @lheagy
Just checking in. Thanks @kbevers, @scivision for creating the issues. @laempy, please keep us posted on your progress addressing these
