### 2018-10-22 41ac232
added "normals" field to LAS
