+++ paper/paper.bib
@@ -1,18 +1,27 @@
-@online{GitHub,
-  author = {GitHub},
-  title = {GitHub.com},
-  year = {2008},
-  url = {https://github.com},
-  urldate = {2018-08-10}
-}
-
 @online{Pyoints_GitHub,
   author = {Sebastian Lamprecht},
   title = {Pyoints},
   year = {2018},
-  url = {https://laempy.github.io/pyoints/modules.html},
-  urldate = {2018-09-27}
+  url = {https://github.com/laempy/pyoints},
+  urldate = {2018-10-22}
 }
+
+@online{Pyoints_PyPI,
+  author = {Sebastian Lamprecht},
+  title = {Pyoints},
+  year = {2018},
+  url = {https://pypi.org/project/pyoints},
+  urldate = {2018-10-22}
+}
+
+@online{Pyoints_docs,
+  author = {Sebastian Lamprecht},
+  title = {Pyoints Package},
+  year = {2018},
+  url = {https://laempy.github.io/pyoints/pyoints.html},
+  urldate = {2018-10-22}
+}
+
 
 @manual{GDAL,
     title = {{GDAL/OGR} Geospatial Data Abstraction software Library},
