+++ paper/paper.bib
@@ -42,13 +42,17 @@
     year = {2000}
 }
 
-@InProceedings{PCL,
-  author    = {Radu Bogdan Rusu and Steve Cousins},
-  title     = {{3D is here: Point Cloud Library (PCL)}},
-  booktitle = {{IEEE International Conference on Robotics and Automation (ICRA)}},
-  month     = {May 9-13},
-  year      = {2011},
-  address   = {Shanghai, China}
+@article{PCL,
+	author = {Rusu, Radu Bogdan and Cousins, Steve},
+	title = {{3D is here: Point Cloud Library (PCL)}},
+	journal = {IEEE International Conference on Robotics and Automation},
+	pages = {1--4},
+	year = {2011},
+	month = {May},
+	issn = {1050-4729},
+	publisher = {IEEE},
+	address   = {Shanghai, China},
+	doi = {10.1109/ICRA.2011.5980567}
 }
 
 @article{Open3D,
