### 2018-10-22 #990 @laempy
Thank you very much for reviewing my package. I have addressed issues 3 to 4 have. I am currently try to solve issue 2. I will push my changes altogether as soon as possible.
