### 2018-10-23 677ee1f
funding information added to README.md
