+++ README.md
@@ -222,6 +222,12 @@ Contribution to the software is always welcome.
 
 # Acknowledgements
 
+## Funding
+
+This work was supported by the [PANTHEON](http://www.project-pantheon.eu) 
+project which is funded by the European Community Horizon 2020 programme 
+under grant agreement 774571.
+
 
 ## Python Depencencies
 
