### 2018-10-23 f91a805
experiments with conda installation
