### 2018-10-29 7fe8c29
LAS field handling revised
