### 2018-10-29 a3998e7
RMSE calculation simplified
