### 2018-10-30 a32c583
experiments with conda
