+++ scripts/conda_install.sh
@@ -1,0 +1,2 @@
+conda env remove --name pyoints_env
+conda create --name pyoints_env pyoints --use-local
