### 2018-10-30 2a9370d
installation instructions adapted
