+++ README.md
@@ -72,80 +72,104 @@ if you like. Thus *Pyoints* is particulary designed for scientists and
 developers of new geo-spatial algorithms.
 
 
+# Installation
 
-# Installation (Ubuntu)
+## Conda installation (recommended)
 
-## Install external dependencies
+Conda installation is recommended. To install *Pyoints* with [Conda]
+(https://conda.io/docs/user-guide/getting-started.html) run:
+
+```
+conda install -c leampy pyoints
+```
+
+
+### Build and install from recipe
+
+To build the package:
+```
+cd /path/to/pyoints/conda
+conda build /path/to/pyoints/conda
+```
+
+Install the builded package:
+```
+conda create --name myenv pyoints --use-local
+```
+
+
+## Manual installation (Debian)
+
+### External dependencies
 
 The following dependencies have to be installed manually, since they are 
-required by the pip packages.
+required by the pip packages. 
 
-### Python
+#### Python
 
-The *Pyoints* targets Python3 >= 3.5.
+*Pyoints* targets Python3 >= 3.5.
 
 
-### Gdal
+#### Gdal
 
 ```
 sudo apt install gdal-bin
 sudo apt install libgdal-dev
 ```
 
-### libspatialindex
+#### libspatialindex
 
 ```
 apt install libspatialindex-dev
 ```
 
-### Liblas
+#### Liblas
 
 ```
 apt install liblas-c3
 ```
 
 
-## Install Pyoints (Ubuntu)
+### Install *Pyoints*
 
 Unfortunately, the gdal version is not detected automatically by `pgdal`.
 Thus, installation with `setup.py` might fail.
 
 
-### Installation via pip (recommended)
+#### Installation via pip (recommended)
 
 ```
 pip install pygdal==$(gdal-config --version).* pyoints
 ```
 
-## Installation with from source
+#### Installation from source
 
 ```
 pip install pygdal==$(gdal-config --version).* -r requirements.txt --upgrade
 ```
 
 
-### Installation with installation script
 
-The installation script asks you to provide a virtual Python environment
-(virtualenv). You need to initialize it first.
+# Development
+
+## Virtualenv (Debian)
+
+To develop the *Pyoints* package `Virtualenv` is recommended. To install the
+ *Pyoints* package conveniently run following command:
+ 
 ```
 ./scripts/pip_install.sh -v path/to/virtualenv
 ```
 
 
-
-# Development
-
-## Virtualenv
-
 ### Install Virtualenv
 
-Installation (Ubuntu)
+Installation
 ```
 apt install virtualenv
 ```
 
-Initialize
+Initialization
 ```
 cd /path/to/library
 virtualenv venv
@@ -153,17 +177,11 @@ virtualenv venv
 
 ### Activate virtualenv
 
-Linux
 ```
 cd /path/to/library
 . venv/bin/activate
 ```
 
-Windows
-```
-cd path\to\library
-venv\Scripts\activate.bat
-```
 
 
 # Documentation
@@ -174,6 +192,7 @@ docstrings with examples for each module and some comprehensive examples. You
 can also take a look at the source files directly.
 
 
+
 # License
 
 Copyright (c) 2018, Sebastian Lamprecht, Trier University,
@@ -181,6 +200,7 @@ lamprecht@uni-trier.de
 
 *Pyoints* is free software made available under the GNU General Public
 License v3 or later (GPLv3+). See [LICENSE](LICENSE) for details.
+
 
 
 # Citing
[…]
