+++ CONTRIBUTING.md
@@ -42,10 +42,64 @@ guidelines.
 
 ### Testing
 
-To test the software please take a look at the [tests](tests) directory.
-Currently doctests are used to test the functionality of a majority of 
-*Pyoints* classes and functions. You can run the file ``tests/test_pyoints.py``
-to run the doctests. After adding a new file, please add a *doctest* reference 
-to ``tests/test_pyoints.py``. To increase the quality of our code we encourage
-you to write additional tests.
+To test the software please take a look at the [tests](pyoints/tests) 
+directory. Currently doctests are used to test the functionality of a majority
+of *Pyoints* classes and functions. You can run the file 
+``tests/test_pyoints.py`` to run the doctests. After adding a new file, please
+add a *doctest* reference to ``tests/test_pyoints.py``. To increase the quality
+of our code we encourage you to write additional tests.
+
+
+
+## Software recommendations
+
+The following Python packages were used for software development, testing and
+documentation.
+
+
+### autopep8
+
+Hideo Hattori
+* [PyPI](https://pypi.org/project/autopep8/)
+* [homepage](https://github.com/hhatto/autopep8)
+* [MIT compatible license](https://github.com/matplotlib/matplotlib/blob/master/LICENSE/LICENSE)
+
+
+### CloudCompare
+
+Daniel Girardeau-Montaut
+* [homepage](https://www.danielgm.net/cc/)
+* [GPL v2](https://github.com/CloudCompare/CloudCompare/blob/master/license.txt)
+
+
+### matplotlib
+
+John D. Hunter, Michael Droettboom
+* [PyPI](https://pypi.org/project/matplotlib/)
+* [homepage](https://matplotlib.org/)
+* [BSD compatible license](https://github.com/matplotlib/matplotlib/blob/master/LICENSE/LICENSE)
+
+
+### pycodestyle
+
+Ian Lee
+* [PyPI](https://pypi.org/project/pycodestyle/)
+* [homepage](https://pycodestyle.readthedocs.io/en/latest/)
+* [Expat license](https://pycodestyle.readthedocs.io/en/latest/index.html#license)
+
+
+### Sphinx
+
+Georg Brandl
+* [PyPI](https://pypi.org/project/Sphinx/)
+* [homepage](http://www.sphinx-doc.org/en/master/)
+* [3-Clause BSD license](https://github.com/sphinx-doc/sphinx)
+
+
+### sphinxcontrib-napoleon
+
+Rob Ruana
+* [PyPI](https://pypi.org/project/sphinxcontrib-napoleon/)
+* [homepage](https://sphinxcontrib-napoleon.readthedocs.io/en/latest/)
+* [2-Clause BSD license](https://github.com/sphinx-contrib/napoleon/blob/master/LICENSE)
 
