### 2018-11-02 6779fef
experiments with conda
