+++ conda/build.sh
@@ -1,1 +1,4 @@
 $PYTHON -m pip install . --ignore-installed -vvv
+# windows
+#python setup.py install --single-version-externally-managed --record=record.txt
+#if errorlevel 1 exit 1
