+++ conda/conda_build_config.yaml
@@ -1,3 +1,3 @@
 python:
     - 3.5
-    - 3.6
+    #- 3.6
