### 2018-11-02 #990 @scivision
@lheagy FYI I am quite tied up on other projects till approx. Nov 15th. If this is an issue someone else can review in the meantime. Sorry about that, this is an interesting project.
