### 2018-11-07 edab576
cross platform tests
