+++ conda/build.sh
@@ -1,4 +1,5 @@
-$PYTHON -m pip install . --ignore-installed -vvv
+#$PYTHON -m pip install . --ignore-installed -vvv
+python -m pip install . --ignore-installed -vvv
 # windows
 #python setup.py install --single-version-externally-managed --record=record.txt
 #if errorlevel 1 exit 1
