+++ conda/conda_build_config.yaml
@@ -1,3 +1,7 @@
 python:
     - 3.5
     #- 3.6
+target_platform:
+    #- win-64
+    #- osx-64
+    #- linux-64
