+++ conda/build.sh
@@ -1,5 +1,11 @@
-#$PYTHON -m pip install . --ignore-installed -vvv
-python -m pip install . --ignore-installed -vvv
+#$PYTHON -m pip install cylinder-fitting plyfile laspy -vvv
+#$PYTHON -m pip install . --no-deps --ignore-installed -vvv
+python -m pip install cylinder-fitting plyfile laspy -vvv
+python -m pip install . --no-deps --ignore-installed -vvv
 # windows
 #python setup.py install --single-version-externally-managed --record=record.txt
+#if errorlevel 1 exit 1
+
+#"%PYTHON%" -m pip install . -vvv
+#python -m pip install . -vvv
 #if errorlevel 1 exit 1
