+++ conda/build.sh
@@ -1,7 +1,11 @@
 #$PYTHON -m pip install cylinder-fitting plyfile laspy -vvv
 #$PYTHON -m pip install . --no-deps --ignore-installed -vvv
-python -m pip install cylinder-fitting plyfile laspy -vvv
-python -m pip install . --no-deps --ignore-installed -vvv
+$PYTHON -m pip install . -vvv
+#python -m pip install . -vvv
+#python -m pip install cylinder-fitting plyfile laspy -vvv
+#python -m pip install . --no-deps --ignore-installed -vvv
+#python -m pip install numpy>=1.15 rtree laspy scipy scikit-learn pyproj dill pandas
+#python -m pip install . -vvv
 # windows
 #python setup.py install --single-version-externally-managed --record=record.txt
 #if errorlevel 1 exit 1
