+++ requirements.txt
@@ -1,10 +1,8 @@
-#pygdal
 numpy>=1.15
 rtree
 laspy
 scipy
 scikit-learn
-pyproj
 dill
 pandas
 opencv-python
