### 2018-11-09 c5c7ddc
conda tests for windows
