+++ requirements.txt
@@ -1,5 +1,4 @@
 numpy>=1.15
-rtree
 laspy
 scipy
 scikit-learn
