+++ conda/conda_build_config.yaml
@@ -2,6 +2,6 @@ python:
     - 3.5
     #- 3.6
 target_platform:
-    #- win-64
+    - win-64
     #- osx-64
     #- linux-64
