### 2018-11-09 779768b
conda tests
