+++ conda/run_test.py
@@ -1,0 +1,4 @@
+# Run tests
+
+print("run unit tests for pyoints")
+from tests import test_pyoints
