### 2018-11-09 ade9c56
cleanup conda
