+++ conda/meta.yaml
@@ -6,88 +6,40 @@ package:
     name: {{ name }}
     version: {{ data.get('version') }}
   
-
 source:
-    #git_url: "https://github.com/laempy/pyoints"
     git_url: "../.git"
   
-
 build:
     number: 0
-    #ignore_prefix_files: True  # required for windows binary files
-    #detect_binary_files_with_prefix: False
-    #script: "{{ PYTHON }} -m pip install . --ignore-installed -vvv"
-
+    skip: True  # [not py3k]
+    
 requirements:
     build:
-        - python #{{ python }}
+        - python
         - pip
         - setuptools
-        #- vc# 14  # [win]
-        #- pyproj
-        #- {{ compiler('c') }}
-        #- {{ compiler('c') }}
     host:
         - python {{ python }}
-        #- vc
-        #- {{ compiler('c') }}
-        #- pip
-        #- setuptools
-        #- {{ compiler('c') }}
-    #dependencies:
-    #    - python {{ python }}
-        #- pip
-        #    - cylinder_fitting
-            #- opencv-python
-        #    - plyfile
-        #    - laspy
     run:
         - python {{ python }}
-        - gdal
-        - pyproj  # pyproj from pip fails for windows
-        - rtree  # rtree from pip fails for windows
-        #- libspatialindex  # throws error
-        #- gdal
-        #- rtree
-        #- python {{ python }}
-        #- gdal
-        #- dill
-        #- opencv
-        #- networkx
-        #- numpy>=1.15
-        #- pandas
-        #- rtree
-        #- scikit-learn
-        #- scipy
-        #- pyproj
-        #- laspy
-        #- plyfile
+        - gdal  # pygdal is not able to detect gdal version
+        - pyproj  # pyproj via pip fails for windows
+        - rtree  # rtree via pip fails for windows
         
-
-
-#test:  # [linux]
 test:
     imports:
         - osgeo
         - cv2
         - laspy
         - plyfile
-        - cylinder_fitting
         - pyproj
         - rtree
+        - cylinder_fitting
         - pyoints
         - pyoints.examples
-        - pyoints.examples.file_examples
         - pyoints.grid
         - pyoints.registration
         - pyoints.storage
-    requirements:
-    #    #- python {{ python }}
-        - unittest
-        - doctest
-        - pkgutil
-    source_files:
-        - tests/test_pyoints.py
         
 about:
     home: https://laempy.github.io/pyoints/modules.html
