### 2018-11-09 39ca354
Installation description revised. fixes #2
