### 2018-11-09 dcd15e9
bugfix and README updated
