+++ README.md
@@ -91,6 +91,8 @@ Currently Linux (64 bit) and Windows (64 bit) are supported.
 If your system is not supported yet, you might build and install it using the 
 [conda reciepe](conda/meta.yaml).
 
+### Build
+
 ```
 conda build /path/to/pyoints/conda 
 ```
@@ -101,10 +103,27 @@ If wished, you can specify the variants, like:
 conda build /path/to/pyoints/conda --variants={python: [3.6], target_platform: ['osx-64']}
 ```
 
-Finally create a virtual environment `myenv` and install *Pyoints*:
+### Installation
+
+Create a virtual environment `myenv` and add the newly build *Pyoints* package:
 
 ```
-conda create -n myenv pyoints --use-local
+conda create -n myenv /path/to/builded/package/pyoints-*-py*_*.tar.bz2 --use-local
+```
+
+Finally install the package:
+
+```
+conda activate myenv
+conda install pyoints
+```
+
+### Test
+
+Test the package by running the provided doctests:
+
+```
+python /path/to/pyoints/tests/test_pyoints.py
 ```
 
 
