### 2018-11-09 6cd71c5
version updated and docs updated
