### 2018-11-09 a6f7929
minor revisions
