+++ scripts/conda_install.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-conda env remove --name pyoints_env
-conda create --name pyoints_env pyoints --use-local
