### 2018-11-10 laempy/pyoints#2 @laempy
@kbevers thank you very much for the suggestion.

A conda recipe has been added and packages have been uploaded to https://anaconda.org/laempy/pyoints. The installation instructions in README.md have been updated accordingly.

Since I have used conda for the first time and some pypi packages were not installed properly, it took a while to build a working recipe. Unfortunately I was not able to cross-compile a osx package. So currently only win-64 and linux-64 packages have been uploaded. I hope the recipe also works for osx.
