### 2018-11-10 laempy/pyoints#5 @laempy
The citations have been revised accordingly.
