### 2018-11-10 laempy/pyoints#3 @laempy
@kbevers thank you for the suggestion.

A CONTRIBUTING.md file with some coding conventions has been added. Also some instructions have been moved from README.md to CONTRIBUTING.md.
