### 2018-11-10 laempy/pyoints#4 @laempy
The DOI has been added.
