### 2018-11-10 #990 @laempy
Hello @Iheagy, I have addressed the issues and published a new version 0.2.0a3 to re-review. I have added some comments on each issue.
Next to some minor revisions I have enabled conda installation as suggested by @kbevers. Unfortunately this took much more time than expected. Currently win-64 and linux-64 packages are provided. The installation instructions have been adapted accordingly. In addition a CONTRIBUTING.md file has been added.
