### 2018-11-10 #990 @kbevers
The changes looks good to me and I have ticked some more boxes above. I most likely won't have time to do a proper test of the code within the next week or two. It does seem to be significantly easier to get started now, good job on that!
