### 2018-12-14 a75a6aa
angle between multiple vectors
