### 2018-12-31 laempy/pyoints#2 @kbevers
The conda recipe didn't work for me on OS X. I have not tried to figure out a way to fix it since I managed to get a working install setup by installing the needed packages with `conda install` and `pip install`, depending on the package. Finally a `pip install -e .` in the source directory installed the package.
