### 2018-12-31 #990 @kbevers
I have now succesfully installed the software, ran the test-suite and confirmed that the package indeed does what it says on the tin. A minor issue in the test suite was found and reporting in https://github.com/laempy/pyoints/issues/7. This is not a blocker for publication but would be nice see fixed in a future release.

As I am working no an unsupported platform (OS X) I did have some trouble installing the package but in the end I managed to make it work.  I have not tested on a supported platform but I have no reason to believe it doesn't work there.
