### 2019-01-03 63d7177
minor bugfixes for issue 7
