### 2019-01-03 f8b7eff
minor bugfixes
