### 2019-01-03 d677604
fixes #7
