### 2019-01-03 e2394b3
bugfix no_data value for rasters
