### 2019-01-04 d3c2b5f
unified round function for docstrings
