+++ pyoints/__init__.py
@@ -56,3 +56,4 @@ from . import (
     transformation,
     vector,
 )
+from .misc import print_rounded

[…]
