### 2019-01-04 24aea62
bugfix raster from_corners
