### 2019-01-04 #990 @laempy
Thanks a lot @kbevers. I have fixed issue laempy/pyoints#7. Since this was just a minor issue, I will push a new version after the review of @scivision.
