### 2019-01-09 laempy/pyoints#8 @lheagy
JOSS review - paper citations
Just a couple minor items with respect to the paper:
- [x] for the Bradski citation, it there a url that could be included to point to the source
- [x] is the consortium, P citation correct? I imagine it is intended to be cited as the "Pantheon Consortium" or similar?
- [x] in the GitHub citation - I would recommend either
  - (preferred): set up [zenodo](http://zenodo.org) with this repository so that the code is archived with a doi and cite that (you will need to create a doi to finalize the JOSS submission anyways)
  - or if you would like to use the github url, then I would suggest that rather than citing the main github url that you cite the link to your repository.

- [x] For the Rusu & Cousins citation: is there a doi for the paper that could be included?

related to openjournals/joss-reviews#990
