### 2019-01-09 laempy/pyoints#9 @lheagy
JOSS review - conda installation error
Related to #2 and openjournals/joss-reviews#990, I was also unsuccessful in trying to install from conda (on my local machine: osx and on binder - e.g. [try jupyter](https://mybinder.org/v2/gh/ipython/ipython-in-depth/master?filepath=binder/Index.ipynb), which runs linux 64bit). Here is the error log I got:

```
Solving environment: failed

CondaHTTPError: HTTP 404 NOT FOUND for url <https://conda.anaconda.org/leampy/noarch/repodata.json>
Elapsed: 00:00.148486
CF-RAY: 4968e3cded3fc56a-ORD

The remote server could not find the noarch directory for the
requested channel with url: https://conda.anaconda.org/leampy

As of conda 4.3, a valid channel must contain a `noarch/repodata.json` and
associated `noarch/repodata.json.bz2` file, even if `noarch/repodata.json` is
empty. please request that the channel administrator create
`noarch/repodata.json` and associated `noarch/repodata.json.bz2` files.
$ mkdir noarch
$ echo '{}' > noarch/repodata.json
$ bzip2 -k noarch/repodata.json

You will need to adjust your conda configuration to proceed.
Use `conda config --show channels` to view your configuration's current state.
Further configuration help can be found at <https://conda.io/docs/config.html>.

```

For the JOSS submission, you do not necessarily have to have this resolved, but I would suggest that in the README, you
- [ ] describe the machine specs for which the conda install should work
- [x] provide alternate installation instructions for installing from source (as @kbevers described in #2):
   - `git clone https://github.com/laempy/pyoints.git`
   - `pip install gdal`  # is there a reason this is not listed in your requirements.txt?
   - `conda install gdal, pyproj, rtree`
   - `pip install -e .`
