### 2019-01-09 laempy/pyoints#10 @lheagy
JOSS Review - examples / getting started
It would be really helpful in the documentation to include a first example to get new users up and running and doing something with `pyoints`. The documentation is really focussed on the API. Having a `Getting Started`,`User Guide`, or `Examples` section with a simple example that produces a plot can help motivate users to explore what they can do with the package. For inspiration, here are a few docs that were recently accepted into JOSS and are well-done:
- https://www.fatiando.org/verde/latest/tutorials/trends.html
- https://jrbourbeau.github.io/pyunfold/examples.html

For the JOSS submission (openjournals/joss-reviews#990), I think clear pointers to a single example from the README or the docs would suffice.
