### 2019-01-09 #990 @lheagy
Hi @laempy: overall, this is looking good! I just have a few comments with respect to the installation and including easy pointers to an example (preferably with a plot) in the README and / or documentation. I have outlined these in issues 8, 9, and 10. Please ping when you have addressed these and I can take another look.
