### 2019-01-10 #990 @laempy
@lheagy thank you very much for these suggestions. I will take a look at the details as soon as possible. Regarding the installation issue: In the README is a typo. `conda install -c laempy pyoints` should work.
