### 2019-01-10 6e26c79
Update of the installation instructions related to issue #9.
