### 2019-01-10 e111084
bibtex revised (see issue #8)
