+++ paper/paper.bib
@@ -31,15 +31,15 @@
     url = {http://gdal.org},
 }
 
-@article{opencv_library,
-    author = {Bradski, G.},
-    citeulike-article-id = {2236121},
-    journal = {Dr. Dobb's Journal of Software Tools},
-    keywords = {bibtex-import},
-    posted-at = {2008-01-15 19:21:54},
-    priority = {4},
-    title = {{The OpenCV Library}},
-    year = {2000}
+@Article{opencv_library,
+  author               = {Bradski, G.},
+  title                = {{The OpenCV Library}},
+  journal              = {Dr. Dobb's Journal of Software Tools},
+  year                 = {2000},
+  url                  = {https://opencv.org},
+  citeulike-article-id = {2236121},
+  keywords             = {bibtex-import},
+  posted-at            = {2008-01-15 19:21:54},
 }
 
 @article{PCL,
@@ -69,12 +69,12 @@
     url = {https://pdal.io/},
 }
 
-@online{PANTHEON,
-  author = {PANTHEON consortium},
-  title = {Precision Farming of Hazelnut Orchards},
-  year = {2018},
-  url = {http://www.project-pantheon.eu/},
-  urldate = {2018-08-10}
+@Online{PANTHEON,
+  author  = {{PANTHEON consortium}},
+  title   = {Precision Farming of Hazelnut Orchards},
+  year    = {2018},
+  url     = {http://www.project-pantheon.eu/},
+  urldate = {2018-08-10},
 }
 
 
