+++ paper/paper.md
@@ -20,57 +20,57 @@ bibliography: paper.bib
 
 # Summary
 
-The evolution of automated systems like autonomous robots and unmanned aerial 
+The evolution of automated systems like autonomous robots and unmanned aerial
 vehicles leads to manifold chances in science, agriculture and industry.
 Remote sensing sensors, like laser scanners and multi-spectral cameras can be
 combined with sensor networks to monitor a research object all-embracingly.
 
-The analysis of such big data is based on techniques of geoinformatics and 
-remote sensing. Next to physically driven approaches, today machine learning 
+The analysis of such big data is based on techniques of geoinformatics and
+remote sensing. Next to physically driven approaches, today machine learning
 techniques are used to extract relevant thematical information the data sets.
 The analysis requires a fusion of the data sets, which is hardened conceptually
-and technically by different data dimensions, data structures and various 
+and technically by different data dimensions, data structures and various
 spatial, spectral and temporal resolutions.
 
 Today various software to deal with these different data sources is available.
-Software like GDAL [@GDAL] and OpenCV [@opencv_library] is intended for image 
+Software like GDAL [@GDAL] and OpenCV [@opencv_library] is intended for image
 processing. Libraries, like PCL [@PCL], Open3D [@Open3D] and PDAL [@PDAL] focus
-on 3D point cloud processing. Each of these software packages provide an API 
-specially designed to solve the problems of their field efficiently. When 
-developing algorithms for automated processing of various types of input data, 
-the differing APIs and programming languages of these software packages become 
-a drawback. To supportfast  algorithm development and a short familiarization, 
+on 3D point cloud processing. Each of these software packages provide an API
+specially designed to solve the problems of their field efficiently. When
+developing algorithms for automated processing of various types of input data,
+the differing APIs and programming languages of these software packages become
+a drawback. To supportfast  algorithm development and a short familiarization,
 a unified API would be desirable.
 
 *Pyoints* is a python package to conveniently process and analyze point
 cloud data, voxels and raster images. It is intended to be used to support
 the development of advanced algorithms for geo-data processing.
 
-The fundamental idea of *Pyoints* is to overcome the conceptual distinction 
-between point clouds, voxel spaces and rasters to simplify data analysis 
-and data fusion of variously structured data. Based on the assumption that any 
-geo-object can be represented by a point, a data structure has been designed 
-which provides a unified API for points, voxels and rasters. Each data 
-structure maintains its characteristic features, to allow for an intuitive use, 
-but all data is also considered as a two or three dimensional point cloud, 
-providing spatial indices which are required in many applications to speed up 
+The fundamental idea of *Pyoints* is to overcome the conceptual distinction
+between point clouds, voxel spaces and rasters to simplify data analysis
+and data fusion of variously structured data. Based on the assumption that any
+geo-object can be represented by a point, a data structure has been designed
+which provides a unified API for points, voxels and rasters. Each data
+structure maintains its characteristic features, to allow for an intuitive use,
+but all data is also considered as a two or three dimensional point cloud,
+providing spatial indices which are required in many applications to speed up
 spatial neighborhood queries.
 
-During development great emphasis was put on designing a powerful but simple 
-API while also providing solutions for most common problems. *Pyoints* 
-implements fundamental functions and some advanced algorithms for point cloud, 
+During development great emphasis was put on designing a powerful but simple
+API while also providing solutions for most common problems. *Pyoints*
+implements fundamental functions and some advanced algorithms for point cloud,
[…]
