### 2019-01-11 da51379
bugfix docstring format
