### 2019-01-11 f76446c
generation of .ipynb notebooks enabled
