+++ sphinx/index.rst
@@ -1,0 +1,8 @@
+Index of Pyoints
+================
+
+.. toctree::
+   :maxdepth: 4
+
+   modules
+   tutorials/getting_started.ipynb
