### 2019-01-17 1b4b2cf
bugfixes coords caching
