+++ sphinx/index.rst
@@ -6,3 +6,5 @@ Index of Pyoints
 
    modules
    tutorials/getting_started.ipynb
+   tutorials/stemdetection.ipynb
+   tutorials/icp.ipynb
