### 2019-01-17 6996f05
minor docstring and pep8 revisions
