### 2019-01-17 0b9e3ad
automatic conversion of jupiter notebooks to .py
