### 2019-01-17 2a0186c
warning loading PLY files fixed
