+++ scripts/generateDocs.sh
@@ -21,7 +21,7 @@ jupyter nbconvert --to script "$TUTORIALS_PATH/*.ipynb"
 cp -r "$SPHINX_PATH" "$COMPILE_PATH"
 cp -r "$TUTORIALS_PATH" "$COMPILE_PATH"
 
-sphinx-apidoc -f -o "$COMPILE_PATH" "$SOURCE_PATH"
+#sphinx-apidoc -f -o "$COMPILE_PATH" "$SOURCE_PATH"
 python3 -m sphinx "$COMPILE_PATH" "$OUT_PATH"
 
 touch "$OUT_PATH/.nojekyll"
