### 2019-01-17 aea6b06
README revised
