+++ scripts/generateDocs.sh
@@ -21,6 +21,11 @@ jupyter nbconvert --to script "$TUTORIALS_PATH/*.ipynb"
 cp -r "$SPHINX_PATH" "$COMPILE_PATH"
 cp -r "$TUTORIALS_PATH" "$COMPILE_PATH"
 
+cp ../README.md "$COMPILE_PATH/"
+cp ../CONTRIBUTING.md "$COMPILE_PATH/"
+cp ../LICENSE "$COMPILE_PATH/"
+
+
 #sphinx-apidoc -f -o "$COMPILE_PATH" "$SOURCE_PATH"
 python3 -m sphinx "$COMPILE_PATH" "$OUT_PATH"
 
