+++ sphinx/index.rst
@@ -1,10 +1,59 @@
-Index of Pyoints
-================
+Pyoints
+=======
+
+
+***************
+Getting Started
+***************
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Contents:
+
+   README.md
+   CONTRIBUTING.md
+   LICENCE.md
+
+
+*********
+Tutorials
+*********
+
+Step by Step
+############
+
+.. toctree::
+   :maxdepth: 4
+
+   tutorials/getting_started.ipynb
+   tutorials/stemdetection.ipynb
+   tutorials/icp.ipynb
+
+
+Downloads
+#########
+
+   :download:`getting_started.py <tutorials/getting_started.py>`
+
+   :download:`getting_started.ipynb <tutorials/getting_started.ipynb>`
+
+
+   :download:`icp.py <tutorials/icp.py>`
+
+   :download:`icp.ipynb <tutorials/icp.ipynb>`
+
+
+   :download:`stemdetection.py <tutorials/stemdetection.py>`
+
+   :download:`stemdetection.ipynb <tutorials/stemdetection.ipynb>`
+
+
+***
+API
+***
 
 .. toctree::
    :maxdepth: 4
 
    modules
-   tutorials/getting_started.ipynb
-   tutorials/stemdetection.ipynb
-   tutorials/icp.ipynb
+
