### 2019-01-18 cad41fe
javascript animation for better compatibility
