### 2019-01-18 af1dfaf
sphinx index with README, CONTRIBUTING and LICENCE
