+++ README.md
@@ -1,7 +1,5 @@
 # Pyoints
 
-![logo](./figures/logo_pyoints.png)
-
 *Pyoints* is a python package to conveniently process and analyze point
 cloud data, voxels and raster images. It is intended to be used to support
 the development of advanced algorithms for geo-data processing.
@@ -72,9 +70,9 @@ if you like. Thus *Pyoints* is particulary designed for scientists and
 developers of new geo-spatial algorithms.
 
 
-# Installation
+## Installation
 
-## Conda Installation
+### Conda Installation
 
 Conda installation is recommended. To install *Pyoints* with
 [Conda](https://conda.io/docs/user-guide/getting-started.html) run:
@@ -86,18 +84,18 @@ conda install -c laempy pyoints
 Currently Linux (64 bit) and Windows (64 bit) have been packaged.
 
 
-## Conda Build
+### Conda Build
 
 If your system is not supported yet, you might build and install it using the
 [conda reciepe](conda/meta.yaml) specifying your desired Python version.
 
-### Build
+#### Build
 
 ```
 conda build /path/to/pyoints/conda --python=3.x
 ```
 
-### Installation
+#### Installation
 
 Create a virtual environment `myenv` and add the newly build *Pyoints* package.
 The package path can be found with `conda build . --output`.
@@ -113,7 +111,7 @@ conda activate myenv
 conda install pyoints
 ```
 
-### Tests
+#### Tests
 
 Test the package by running the provided doctests:
 
@@ -122,7 +120,7 @@ python /path/to/pyoints/tests/test_pyoints.py
 ```
 
 
-## Installation from source
+### Installation from source
 
 We recommend to use a virtual environment (like Conda) to install *Pyoints*.
 
@@ -145,7 +143,7 @@ Unfortunately, the gdal version is not detected automatically by *pgdal*. Thus,
 we have to specify the gdal version expicilely with ```gdal-config --version```.
 
 
-# Documentation
+## Documentation
 
 You can find the documentation at
 [docs](https://laempy.github.io/pyoints/pyoints.html). Here you can find
@@ -154,7 +152,7 @@ can also take a look at the source files directly.
 
 
 
-# License
+## License
 
 Copyright (c) 2018, Sebastian Lamprecht, Trier University,
 lamprecht@uni-trier.de
@@ -164,7 +162,7 @@ License v3 or later (GPLv3+). See [LICENSE](LICENSE) for details.
 
 
 
-# Citing
+## Citing
 
 Please cite `Pyoints` as:
 
@@ -185,22 +183,22 @@ URL [https://github.com/laempy/pyoints](https://github.com/laempy/pyoints),
 ```
 
 
-# Contributing
+## Contributing
 
 Any conribution to the *pyoints* project is welcome. See
 [CONTRIBUTING](CONTRIBUTING.md) for details.
 
 
 
-# Acknowledgements
+## Acknowledgements
 
-## Funding
+### Funding
 
 This work was supported by the [PANTHEON](http://www.project-pantheon.eu)
 project which is funded by the European Community Horizon 2020 programme
 under grant agreement 774571.
 
-## Datasets
+### Datasets
 
 For the tutorials the "Stanford Bunny" dataset has been used:
 
[…]
