+++ sphinx/index.rst
@@ -2,50 +2,42 @@ Pyoints
 =======
 
 
-***************
-Getting Started
-***************
+********
+Overview
+********
+
 
 .. toctree::
-   :maxdepth: 1
-   :caption: Contents:
+   :maxdepth: 2
 
-   README.md
-   CONTRIBUTING.md
-   LICENCE.md
+   readme
+   contributing
+
+.. toctree::
+   :titlesonly:
+   
+   license
 
 
 *********
 Tutorials
 *********
 
-Step by Step
-############
-
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 2
 
    tutorials/getting_started.ipynb
    tutorials/stemdetection.ipynb
    tutorials/icp.ipynb
 
 
-Downloads
-#########
+.. csv-table:: Downloads
+   :header: "Topic", "Jupiter Notebook", "Python Script"
+   :widths: 20, 20, 20
 
-   :download:`getting_started.py <tutorials/getting_started.py>`
-
-   :download:`getting_started.ipynb <tutorials/getting_started.ipynb>`
-
-
-   :download:`icp.py <tutorials/icp.py>`
-
-   :download:`icp.ipynb <tutorials/icp.ipynb>`
-
-
-   :download:`stemdetection.py <tutorials/stemdetection.py>`
-
-   :download:`stemdetection.ipynb <tutorials/stemdetection.ipynb>`
+   "Getting started", :download:`getting_started.ipynb <tutorials/getting_started.ipynb>`, :download:`getting_started.py <tutorials/getting_started.py>`
+   "ICP", :download:`icp.ipynb <tutorials/icp.ipynb>`, :download:`icp.py <tutorials/icp.py>`
+   "Stem detection", :download:`stemdetection.ipynb <tutorials/stemdetection.ipynb>`, :download:`stemdetection.py <tutorials/stemdetection.py>`
 
 
 ***
@@ -56,4 +48,8 @@ API
    :maxdepth: 4
 
    modules
+
+
+
+
 
