### 2019-01-18 ea918e2
bugfix ply dtype
