### 2019-01-18 ff7bee6
stem detection tutorial renamed
