+++ sphinx/index.rst
@@ -27,7 +27,7 @@ Tutorials
    :maxdepth: 2
 
    tutorials/getting_started.ipynb
-   tutorials/stemdetection.ipynb
+   tutorials/stem_detection.ipynb
    tutorials/icp.ipynb
 
 
@@ -37,7 +37,7 @@ Tutorials
 
    "Getting started", :download:`getting_started.ipynb <tutorials/getting_started.ipynb>`, :download:`getting_started.py <tutorials/getting_started.py>`
    "ICP", :download:`icp.ipynb <tutorials/icp.ipynb>`, :download:`icp.py <tutorials/icp.py>`
-   "Stem detection", :download:`stemdetection.ipynb <tutorials/stemdetection.ipynb>`, :download:`stemdetection.py <tutorials/stemdetection.py>`
+   "Stem detection", :download:`stem_detection.ipynb <tutorials/stem_detection.ipynb>`, :download:`stem_detection.py <tutorials/stem_detection.py>`
 
 
 ***
