+++ scripts/generateDocs.sh
@@ -21,13 +21,9 @@ fi
 cp -r "$SPHINX_PATH" "$COMPILE_PATH"
 cp -r "$TUTORIALS_PATH" "$COMPILE_PATH"
 
-cp ../README.md "$COMPILE_PATH/"
-cp ../CONTRIBUTING.md "$COMPILE_PATH/"
-cp ../LICENSE "$COMPILE_PATH/"
-
 jupyter nbconvert --to script "$COMPILE_PATH/*.ipynb"
 
-#sphinx-apidoc -f -o "$COMPILE_PATH" "$SOURCE_PATH"
+sphinx-apidoc -f -o "$COMPILE_PATH" "$SOURCE_PATH"
 python3 -m sphinx "$COMPILE_PATH" "$OUT_PATH"
 
 touch "$OUT_PATH/.nojekyll"
