### 2019-01-18 b2cf589
updates on installation
