### 2019-01-18 aa2ccac
bugfix docs generation
