+++ docs/_downloads/67f17cf78199fffcda57740e2bff9808/stem_detection.py
@@ -1,0 +1,359 @@
+
+# coding: utf-8
+
+# # Detection of tree stems with Pyoints
+# In the following, we try to detect stems in a forest using a three dimensional point cloud generated by a terrestrial laser scanner.
+# The basic steps are:
+# 
+# 	1. Loading of the point cloud data
+# 	2. Calculation of the height above ground
+# 	3. Fitering of stem points
+# 	4. Identification of stem clusters
+# 	5. Fitting of stem vectors
+
+# We begin with loading the required modules.
+
+# In[ ]:
+
+
+import numpy as np
+
+from pyoints import (
+	storage,
+	Extent,
+	filters,
+	interpolate,
+	clustering,
+	classification,
+	vector,
+)
+
+
+# In[ ]:
+
+
+from mpl_toolkits import mplot3d
+import matplotlib.pyplot as plt 
+
+get_ipython().run_line_magic('matplotlib', 'inline')
+
+
+# ## Loading of the point cloud data
+
+# We load a three dimensional point cloud scanned in a forest.
+
+# In[ ]:
+
+
+lasReader = storage.LasReader('forest.las')
+las = lasReader.load()
+print(len(las))
+
+
+# We recieve a LasRecords epresenting the point cloud. We can treat it like a numpy record array with additional features. So we inspect its properties first.
+
+# In[ ]:
+
+
+print('shape:')
+print(las.shape)
+print('attributes:')
+print(las.dtype.descr)
+print('projection:')
+print(las.proj.proj4)
+print('transformation matrix:')
+print(np.round(las.t, 2))
+print('data:')
+print(las)
+
+
+# ## Calculation of the height above ground
+# The LAS file provides some altitude values. But we are interrested in the height above ground instead. So, we select points representing the ground first, to fit a digital elevation model (DEM). Using the DEM we calculate the height above ground for each point.
+
+# We use a DEM filter with a resolution of 0.5m. The filter selects local minima and guarantees a horizontal point distance of at least 0.5m and an altitude change between neighbored points below 50 degree.
+
+# In[ ]:
+
+
+grd_ids = filters.dem_filter(las.coords, 0.5, max_angle=50)
+print(grd_ids)
+
+
+# We have recieved a list of point indices which is used to select the desired representative ground points. So, we update the classification field for these points and plotthem.
+
+# In[ ]:
+
+
+las.classification[grd_ids] = 22
+
+
+# In[ ]:
+
+
+plt.scatter(*las[las.classification == 22].coords[:, :2].T, color='black')
+plt.xlabel('X (m)')
+plt.ylabel('Y (m)')
+plt.show()
+
+
+# Now we fit a DEM using a nearest neighbor interpolator.
+
+# In[ ]:
+
+
+xy = las.coords[grd_ids, :2]
+z = las.coords[grd_ids, 2]
+dem = interpolate.KnnInterpolator(xy, z)
+
+
+# Finally we calculate the height above ground. And add a corresponding height attribute to the point cloud.
+
+# In[ ]:
+
+
+height = las.coords[:, 2] - dem(las.coords)
+las = las.add_fields([('height', float)], data=[height])
+print(las.dtype.descr)
+
+
+# ## Fitering of stem points
[…]
