+++ docs/_downloads/cd3fd2c207914aec0e3b57509e5b1a4e/getting_started.py
@@ -1,0 +1,607 @@
+
+# coding: utf-8
+
+# # Getting started
+# In this tutorial we will learn the basics of raster and point cloud processing using Pyoints.
+
+# We begin with loading the required modules.
+
+# In[ ]:
+
+
+import numpy as np
+
+from pyoints import (
+	nptools,
+	Proj,
+	GeoRecords,
+	Grid,
+	Extent,
+	transformation,
+	filters,
+	clustering,
+	classification,
+	smoothing,
+)
+
+
+# In[ ]:
+
+
+from mpl_toolkits import mplot3d
+import matplotlib.pyplot as plt
+
+get_ipython().run_line_magic('matplotlib', 'inline')
+
+
+# ## Grid
+# We create a two dimensional raster by providing a projection system, a transformation matrix and some data. The transformation matrix defines the origin and scale of the raster. For this example, we just use the default projection system. 
+
+# In[ ]:
+
+
+print('projection')
+proj = Proj()
+print(proj)
+
+print('numpy record array:')
+rec = nptools.recarray(
+	{'indices': np.mgrid[0:50, 0:30].T},
+	dim=2
+)
+print(rec.shape)
+print(rec.dtype.descr)
+
+print('transformation matrix')
+T = transformation.matrix(t=[-15, 10], s=[0.8, -0.8])
+print(T)
+
+
+# In[ ]:
+
+
+grid = Grid(proj, rec, T)
+
+
+# Let's inspect the properties of the raster.
+
+# In[ ]:
+
+
+print('shape:')
+print(grid.shape)
+print('number of cells:')
+print(grid.count)
+print('fields:')
+print(grid.dtype)
+print('projection:')
+print(grid.proj)
+print('transformation matrix:')
+print(np.round(grid.t, 2))
+print('origin:')
+print(np.round(grid.t.origin, 2).tolist())
+print('extent:')
+print(np.round(grid.extent(), 2).tolist())
+
+
+# Now, we visuallize the x 'indices' of the raster.
+
+# In[ ]:
+
+
+fig = plt.figure(figsize=(10, 10))
+plt.xlabel('X (pixels)')
+plt.ylabel('Y (pixels)')
+
+plt.imshow(grid.indices[:, :, 0])
+plt.show()
+
+
+# You might have noticed, that the field 'coords' has been implicitly added to the record array. The coordinates correspond to the centers of the raster cells.
+
+# In[ ]:
+
+
+print(np.round(grid.coords, 2))
+
+
+# Based on these coordinates we create an additional fieldrepresenting a surface.
+
+# In[ ]:
+
+
+x = grid.coords[:, :, 0]
+y = grid.coords[:, :, 1]
+dist = np.sqrt(x ** 2 + y ** 2)
+z = 9 + 10 * (np.sin(0.5 * x) / np.sqrt(dist + 1) + np.cos(0.5 * y) / np.sqrt(dist + 1))
+grid = grid.add_fields([('z', float)], data=[z])
+print(grid.dtype.descr)
+
+
+# In[ ]:
+
+
+fig = plt.figure(figsize=(10, 10))
+plt.xlabel('X (pixels)')
+plt.ylabel('Y (pixels)')
+
+plt.imshow(grid.z, cmap='coolwarm')
+plt.show()
+
+
+# If we like to treat the raster as a point cloud or list of points, we call the 'records' function. As a result we recieve a flattened version of the raster.
+
+# In[ ]:
+
+
[…]
