+++ docs/_sources/tutorials/icp.ipynb.txt
@@ -1,0 +1,580 @@
+{
+    "cells":[
+        {
+            "cell_type":"markdown",
+            "metadata":{},
+            "source":[
+                "# ICP for point cloud alignment",
+                "\n",
+                "In this tutorial we will learn to align several point clouds",
+                "using two variants of the Iterative Closest Point (ICP) ",
+                "algorithm."
+            ]
+        },
+        {
+            "cell_type":"markdown",
+            "metadata":{},
+            "source":[
+                "We begin with loading the required modules."
+            ]
+        },
+        {
+            "cell_type":"code",
+            "execution_count":null,
+            "metadata":{},
+            "outputs":[],
+            "source":[
+                "import numpy as np\n",
+                "\n",
+                "from pyoints import (\n",
+                "\tstorage,\n",
+                "\tExtent,\n",
+                "\ttransformation,\n",
+                "\tfilters,\n",
+                "\tregistration,\n",
+                "\tnormals,\n",
+                ")"
+            ]
+        },
+        {
+            "cell_type":"code",
+            "execution_count":null,
+            "metadata":{},
+            "outputs":[
+
+            ],
+            "source":[
+                "from mpl_toolkits import mplot3d\n",
+                "import matplotlib.pyplot as plt\n",
+                "from matplotlib import animation\n",
+                "from IPython.display import HTML\n",
+                "\n",
+                "%matplotlib inline"
+            ]
+        },
+        {
+            "cell_type":"markdown",
+            "metadata":{},
+            "source":[
+                "## Preparing the data\n",
+                "We load three point clouds of the ",
+                "Standfort Bunny [[1]](#References) dataset. The ",
+                "original PLY files have been converted to a binary format ",
+                "to spare disc space and speed up loading time."
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": null,
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "A = storage.loadPly('bun000_binary.ply')\n",
+                "print(A.shape)\n",
+                "print(A.dtype.descr)"
+            ]
+        },
+                {
+            "cell_type": "code",
+            "execution_count": null,
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "B = storage.loadPly('bun045_binary.ply')\n",
+                "print(B.shape)\n",
+                "print(B.dtype.descr)"
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": null,
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "C = storage.loadPly('bun090_binary.ply')\n",
+                "print(C.shape)\n",
+                "print(C.dtype.descr)"
+            ]
+        },
+        {
+            "cell_type":"markdown",
+            "metadata":{},
+            "source":[
+                "We filter the point cloud to recieve sparse point clouds ",
+                "more suitable for visualization. Thinning the point clouds ",
+                "also speeds up the ICP algorithm."
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": null,
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "r = 0.004\n",
+                "A = A[list(filters.ball(A.indexKD(), r))]\n",
+                "B = B[list(filters.ball(B.indexKD(), r))]\n",
+                "C = C[list(filters.ball(C.indexKD(), r))]"
+            ]
+        },
+        {
+            "cell_type":"markdown",
+            "metadata":{},
[…]

[…]
