### 2019-01-18 3c263cf
new structure added to update License
