### 2019-01-21 7f77271
numpy requirement removed for conda tests
