+++ requirements.txt
@@ -1,4 +1,3 @@
-numpy>=1.15
 laspy
 scipy
 scikit-learn
