+++ requirements.txt
@@ -1,3 +1,4 @@
+numpy>=1.15
 laspy
 scipy
 scikit-learn
