### 2019-01-21 f4f4fec
gdal installation problems test
