+++ requirements.txt
@@ -1,4 +1,5 @@
 numpy>=1.15
+gdal
 laspy
 scipy
 scikit-learn
