+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy>=1.15
+#numpy>=1.15
 #gdal
 laspy
 scipy
