+++ conda/meta.yaml
@@ -20,10 +20,11 @@ requirements:
         - setuptools
     host:
         - python {{ python }}
-        - GDAL  # pygdal is not able to detect gdal version
-        - cython
+        #- GDAL  # pygdal is not able to detect gdal version
     run:
         - python {{ python }}
+        - gdal  # pygdal is not able to detect gdal version
+        - numpy>=1.15
         #- gdal  # pygdal is not able to detect gdal version
         - pyproj  # pyproj via pip fails for Windows
         - rtree  # rtree via pip fails for Windows
