+++ conda/meta.yaml
@@ -21,6 +21,8 @@ requirements:
     host:
         - python {{ python }}
         #- GDAL  # pygdal is not able to detect gdal version
+        - gdal  # pygdal is not able to detect gdal version
+        - numpy>=1.15
     run:
         - python {{ python }}
         - gdal  # pygdal is not able to detect gdal version
