+++ requirements.txt
@@ -1,9 +1,10 @@
-#numpy>=1.15
-#gdal
+numpy>=1.15
+gdal
 laspy
 scipy
 scikit-learn
 dill
+pandas
 opencv-python
 plyfile
 networkx
