+++ conda/meta.yaml
@@ -24,8 +24,6 @@ requirements:
     run:
         - python {{ python }}
         - gdal  # pygdal is not able to detect gdal version
-        - numpy>=1.15
-        - pandas
         #- gdal  # pygdal is not able to detect gdal version
         - pyproj  # pyproj via pip fails for Windows
         - rtree  # rtree via pip fails for Windows
