+++ conda/meta.yaml
@@ -24,6 +24,7 @@ requirements:
     run:
         - python {{ python }}
         - gdal  # pygdal is not able to detect gdal version
+        - numpy>=1.15
         #- gdal  # pygdal is not able to detect gdal version
         - pyproj  # pyproj via pip fails for Windows
         - rtree  # rtree via pip fails for Windows
