+++ conda/meta.yaml
@@ -1,6 +1,6 @@
 {% set name = "pyoints" %}
 {% set data = load_setup_py_data() %}
-
+{% set numpy_min_version = "1.15" %}
 
 package:
     name: {{ name }}
@@ -20,14 +20,12 @@ requirements:
         - setuptools
     host:
         - python {{ python }}
-        #- GDAL  # pygdal is not able to detect gdal version
         - gdal  # pygdal is not able to detect gdal version
         - numpy>=1.15
     run:
         - python {{ python }}
-        - gdal  # pygdal is not able to detect gdal version
-        - numpy>=1.15
-        #- gdal  # pygdal is not able to detect gdal version
+        - {{ pin_compatible('gdal') }}
+        - {{ pin_compatible('numpy') }}
         - pyproj  # pyproj via pip fails for Windows
         - rtree  # rtree via pip fails for Windows
 
