### 2019-02-04 eee678a
cleanup installation
