+++ conda/conda_build_config.yaml
@@ -1,4 +1,4 @@
 python:
-  #- 3.5
-  #- 3.6
+  - 3.5
+  - 3.6
   - 3.7
