### 2019-02-05 b6fe216
conda script revised
