+++ conda/meta.yaml
@@ -15,7 +15,7 @@ build:
 
 requirements:
     build:
-        - python
+        - python {{ python }}
         - pip
         - setuptools
     host:
