+++ docs/_downloads/51f8ada99bd2302aff2952eb3c048c50/icp.py
@@ -1,351 +1,0 @@
-
-# coding: utf-8
-
-# # ICP for point cloud alignment
-# In this tutorial we will learn to align several point cloudsusing two variants of the Iterative Closest Point (ICP) algorithm.
-
-# We begin with loading the required modules.
-
-# In[ ]:
-
-
-import numpy as np
-
-from pyoints import (
-	storage,
-	Extent,
-	transformation,
-	filters,
-	registration,
-	normals,
-)
-
-
-# In[ ]:
-
-
-from mpl_toolkits import mplot3d
-import matplotlib.pyplot as plt
-from matplotlib import animation
-from IPython.display import HTML
-
-get_ipython().run_line_magic('matplotlib', 'inline')
-
-
-# ## Preparing the data
-# We load three point clouds of the Standfort Bunny [[1]](#References) dataset. The original PLY files have been converted to a binary format to spare disc space and speed up loading time.
-
-# In[ ]:
-
-
-A = storage.loadPly('bun000_binary.ply')
-print(A.shape)
-print(A.dtype.descr)
-
-
-# In[ ]:
-
-
-B = storage.loadPly('bun045_binary.ply')
-print(B.shape)
-print(B.dtype.descr)
-
-
-# In[ ]:
-
-
-C = storage.loadPly('bun090_binary.ply')
-print(C.shape)
-print(C.dtype.descr)
-
-
-# We filter the point cloud to recieve sparse point clouds more suitable for visualization. Thinning the point clouds also speeds up the ICP algorithm.
-
-# In[ ]:
-
-
-r = 0.004
-A = A[list(filters.ball(A.indexKD(), r))]
-B = B[list(filters.ball(B.indexKD(), r))]
-C = C[list(filters.ball(C.indexKD(), r))]
-
-
-# Before we visuallize the point cloud, we define the colors and the axes limits.
-
-# In[ ]:
-
-
-axes_lims = Extent([
-	A.extent().center - 0.5 * A.extent().ranges.max(),
-	A.extent().center + 0.5 * A.extent().ranges.max()
-])
-colors = {'A': 'green', 'B': 'blue', 'C': 'red'}
-
-
-# In[ ]:
-
-
-fig = plt.figure(figsize=(15, 15))
-ax = plt.axes(projection='3d')
-ax.set_xlim(axes_lims[0], axes_lims[3])
-ax.set_ylim(axes_lims[1], axes_lims[4])
-ax.set_zlim(axes_lims[2], axes_lims[5])
-ax.set_xlabel('X (m)')
-ax.set_ylabel('Y (m)')
-ax.set_zlabel('Z (m)')
-
-ax.scatter(*A.coords.T, color=colors['A'])
-ax.scatter(*B.coords.T, color=colors['B'])
-ax.scatter(*C.coords.T, color=colors['C'])
-plt.show()
-
-
-# We can see, that the point clouds B and C are rotated by 45 and 90 degree. Since the ICP algorithm assumes already roughly aligned point clouds as an input, we rotate the point clouds accordingly. But to harden the problem a bit, we use slightly differing rotation angles.
-
-# In[ ]:
-
-
-T_A = transformation.r_matrix([90*np.pi/180, 0, 0])
-A.transform(T_A)
-T_B = transformation.r_matrix([86*np.pi/180, 0, 45*np.pi/180])
-B.transform(T_B)
-T_C = transformation.r_matrix([95*np.pi/180, 0, 90*np.pi/180])
-C.transform(T_C)
-
-
-# We update axes limits and visualize the point clouds again. The resulting point clouds serve as input for the ICP algorithm.
-
-# In[ ]:
-
-
-axes_lims = Extent([
[…]
