### 2019-02-05 3fb6cf1
Update of references and version info
