+++ paper/paper.bib
@@ -1,27 +1,17 @@
 @online{Pyoints_GitHub,
   author = {Sebastian Lamprecht},
   title = {Pyoints},
-  year = {2018},
-  url = {https://github.com/laempy/pyoints},
-  urldate = {2018-10-22}
-}
-
-@online{Pyoints_PyPI,
-  author = {Sebastian Lamprecht},
-  title = {Pyoints},
-  year = {2018},
-  url = {https://pypi.org/project/pyoints},
-  urldate = {2018-10-22}
+  year = {2019},
+  doi = {10.5281/zenodo.2557575}
 }
 
 @online{Pyoints_docs,
   author = {Sebastian Lamprecht},
-  title = {Pyoints Package},
-  year = {2018},
+  title = {Pyoints},
+  year = {2019},
   url = {https://laempy.github.io/pyoints/pyoints.html},
-  urldate = {2018-10-22}
+  urldate = {2019-02-05}
 }
-
 
 @manual{GDAL,
     title = {{GDAL/OGR} Geospatial Data Abstraction software Library},
