+++ README.md
@@ -166,7 +166,7 @@ License v3 or later (GPLv3+). See [LICENSE](LICENSE) for details.
 
 ## Citing
 
-Please cite `Pyoints` as:
+Please cite *Pyoints* as follows:
 
 S. Lamprecht (2019):
 "Pyoints: A Python package for point cloud, voxel and raster processing". [![DOI](https://zenodo.org/badge/144562205.svg)](https://zenodo.org/badge/latestdoi/144562205)
