### 2019-02-05 laempy/pyoints#10 @laempy
@lheagy, thank you for this suggestion. I have added three tutorials as jupiter notebooks. The docs have been revised accordingly. A link from the README to the docs is given.
