### 2019-02-05 laempy/pyoints#9 @laempy
Thank you @lheagy. The installation failed because of a typo in the installation instructions. Many thanks to @ReedAnders, who has noticed it. I have also added alternative installation instructions to the README file.

Gdal, pyproj, rtree and numpy are not included in the requirements file, because of version clashes during conda installation (pygdal does not recognize the available gdal versions).

Unfortunately I have no glue which machine specs are required. Packages for Windows (64 bit) and Linux (64 bit) have been prepared with conda. I assume, also 32 bit and Mac should work, but in the moment I have no chance to test it. Thus, also a conda recipe is provided to be able to compile a package on your own.
