### 2019-02-05 laempy/pyoints#7 @laempy
Thank you @kbevers. To get rid of the numerical instabilities I have added functions to round the output within the docstrings with a given tolerance. Thus, the tests should work now, even if a slightly different machine precision is used.
