### 2019-02-11 #990 @laempy
Hi @lheagy. I have addressed the open issues and created a new version v0.2.0b1. Could you please take a second look? Unfortunately I wasn't able to create a new conda package yet, but I will do that as soon as I find enough time. In consequence, currently the package needs to be installed from source according to the README.
