### 2019-02-24 laempy/pyoints#8 @lheagy
Looks good, thanks @laempy
