### 2019-02-24 laempy/pyoints#8 @lheagy
It looks like there are doi's for The OpenCV Library and Open3D
```
MISSING DOIs

- https://doi.org/10.1109/mra.2009.933612 may be missing for title: The OpenCV Library
- https://doi.org/10.1109/cvpr.2018.00409 may be missing for title: Open3D: A Modern Library for 3D Data Processing
```

Would you mind double checking these and if they are appropriate adding them to the paper?
