### 2019-02-24 laempy/pyoints#9 @lheagy
Thanks for clarifying @laempy, this looks good.
