### 2019-02-24 #990 @lheagy
👋 Hi @laempy, thanks for your changes - I really like the tutorials you added. Just one outstanding item from my perspective, that is the doi's mentioned above (laempy/pyoints#8). Would you mind taking a look and see if those are appropriate to add to the paper?

@kbevers, @scivision, thanks for your efforts on the review. Would you like to take a look at the revisions and add any final comments?
