+++ pyoints/storage/LasHandler.py
@@ -22,6 +22,7 @@
 import os
 import numpy as np
 import laspy
+import liblas
 
 from ..extent import Extent
 from ..georecords import (
@@ -37,6 +38,9 @@ from .. import (
 
 from .BaseGeoHandler import GeoFile
 from .dtype_converters import numpy_to_laspy_dtype
+
+
+SUPPORTED_FORMATS = [0, 1, 2, 3, 4, 5]
 
 
 class LasReader(GeoFile):
@@ -61,13 +65,14 @@ class LasReader(GeoFile):
 
         lasFile = laspy.file.File(self.file, mode='r')
 
-        # try to read projection from WKT
+        # try to read projection from file
         if proj is None:
-            for vlr in lasFile.header.vlrs:
-                if vlr.record_id == 2112:
-                    wkt = str(vlr.VLR_body.decode('utf-8'))
-                    proj = projection.Proj.from_wkt(wkt)
-                    break
+            reader = liblas.file.File(self.file, mode='r')
+            wkt = reader.header.srs.get_wkt().decode('utf-8')
+            if not wkt == '':
+                proj = projection.Proj.from_wkt(wkt)
+            reader.close()
+
         self.proj = proj
 
         self.t = transformation.t_matrix(lasFile.header.offset)
@@ -93,12 +98,12 @@ class LasReader(GeoFile):
         lasFile = laspy.file.File(self.file, mode='r')
 
         # check point formats
-        supported_formats = [0, 1, 2, 3, 4, 5]
-        if lasFile.header.data_format_id not in supported_formats:
+
+        if lasFile.header.data_format_id not in SUPPORTED_FORMATS:
             m = "Only point formats %s supported yet, got %"
             raise ValueError(
                 m %
-                (supported_formats, lasFile.header.data_format_id))
+                (SUPPORTED_FORMATS, lasFile.header.data_format_id))
 
         scale = np.array(lasFile.header.scale, dtype=np.float64)
         offset = np.array(lasFile.header.offset, dtype=np.float64)
@@ -175,7 +180,7 @@ class LasReader(GeoFile):
         return LasRecords(self.proj, data, T=t)
 
 
-def writeLas(geoRecords, outfile):
+def writeLas(geoRecords, outfile, point_format=3):
     """ Write a LAS file to disc.
 
     Parameters
@@ -184,6 +189,8 @@ def writeLas(geoRecords, outfile):
         Points to store to disk.
     outfile : String
         Desired output file.
+    point_format : optional, positive int
+        Desired LAS point format. See LAS specification for details.
 
     """
     # validate input
@@ -192,28 +199,37 @@ def writeLas(geoRecords, outfile):
     if not os.access(os.path.dirname(outfile), os.W_OK):
         raise IOError('File %s is not writable' % outfile)
 
+    if point_format not in SUPPORTED_FORMATS:
+        raise ValueError("'point_format' %s not supported" % str(point_format))
+
     records = geoRecords.records()
 
     # Create file header
-    header = laspy.header.Header(file_version=1.3, point_format=3)
+    header = laspy.header.Header(file_version=1.3, point_format=point_format)
     header.file_sig = 'LASF'
-    #header.format = 1.3
-    #header.data_format_id = 7
 
     # Open file in write mode
     lasFile = laspy.file.File(outfile, mode='w', header=header)
 
-    # create projection VLR using WKT
-    proj_vlr = laspy.header.VLR(
-        user_id="LASF_Projection",
-        record_id=2112,
-        VLR_body=str.encode(geoRecords.proj.wkt),
-        description="OGC Coordinate System WKT"
[…]
