+++ requirements.txt
@@ -3,6 +3,7 @@
 #rtree
 #pyproj
 laspy
+liblas
 scipy
 scikit-learn
 dill
