### 2019-02-27 #990 @lheagy
👋 @laempy: could you please make a new release, archive the software on [zenodo](http://zenodo.org) or similar and post the doi here?
