### 2019-02-27 137cffb
improved handling of missing projection
