### 2019-02-27 47ee160
test printing within docstings
