+++ pyoints/transformation.py
@@ -41,6 +41,7 @@ from .assertion import (
     isnumeric,
 )
 from .misc import print_rounded
+print_rounded(1)
 
 
 def transform(coords, T, inverse=False, precise=False):
@@ -357,7 +358,7 @@ def r_matrix(a, order='xyz'):
     Two dimensional case.
 
     >>> R = r_matrix(np.pi/4)
-    >>> print_rounded(R, 3)
+    >>> print(R, 3)
     [[ 0.707 -0.707  0.   ]
      [ 0.707  0.707  0.   ]
      [ 0.     0.     1.   ]]
@@ -365,7 +366,7 @@ def r_matrix(a, order='xyz'):
     Three dimensional case.
 
     >>> R = r_matrix([np.pi/2, 0, np.pi/4])
-    >>> print_rounded(R, 3)
+    >>> print(R, 3)
     [[ 0.707  0.     0.707  0.   ]
      [ 0.707  0.    -0.707  0.   ]
      [ 0.     1.     0.     0.   ]
@@ -446,7 +447,7 @@ def q_matrix(q):
     --------
 
     >>> T = q_matrix([0.7071, 0.7071, 0, 0])
-    >>> print_rounded(T, 2)
+    >>> print(T, 2)
     [[ 1.  0.  0.  0.]
      [ 0.  0. -1.  0.]
      [ 0.  1.  0.  0.]
@@ -491,12 +492,12 @@ def add_dim(T):
     Two dimensional case.
 
     >>> T = matrix(t=[2, 3 ], s=[0.5, 2])
-    >>> print_rounded(T, 3)
+    >>> print(T, 3)
     [[0.5 0.  2. ]
      [0.  2.  3. ]
      [0.  0.  1. ]]
     >>> T = add_dim(T)
-    >>> print_rounded(T, 3)
+    >>> print(T, 3)
     [[0.5 0.  0.  2. ]
      [0.  2.  0.  3. ]
      [0.  0.  1.  0. ]
@@ -691,7 +692,7 @@ class LocalSystem(np.ndarray, object):
      [ 0.   0.   1. ]]
     >>> print(L.dim)
     2
-    >>> print_rounded(L.origin)
+    >>> print(L.origin)
     [2. 3.]
 
     See Also
@@ -812,7 +813,7 @@ class LocalSystem(np.ndarray, object):
 
         >>> T = matrix(t=[2, 3], s=[0.5, 10])
         >>> lcoords = T.to_local([(0, 0), (0, 1), (1, 0), (-1, -1)])
-        >>> print_rounded(lcoords)
+        >>> print(lcoords)
         [[ 2.   3. ]
          [ 2.  13. ]
          [ 2.5  3. ]
@@ -840,7 +841,7 @@ class LocalSystem(np.ndarray, object):
 
         >>> T = matrix(t=[2, 3], s=[0.5, 10])
         >>> gcoords = T.to_global([(2, 3), (2, 13), (2.5, 3), (1.5, -7)])
-        >>> print_rounded(gcoords)
+        >>> print(gcoords)
         [[ 0.  0.]
          [ 0.  1.]
          [ 1.  0.]
@@ -873,7 +874,7 @@ class LocalSystem(np.ndarray, object):
          [0. 2. 0.]
          [0. 0. 1.]]
         >>> e = T.explained_variance([(2, 1), (0, 0), (1, 1), (2, 3)])
-        >>> print_rounded(e, 3)
+        >>> print(e, 3)
         [0.688 4.75 ]
 
         See Also
@@ -908,7 +909,7 @@ class LocalSystem(np.ndarray, object):
          [0. 2. 0.]
          [0. 0. 1.]]
[…]
