### 2019-02-27 4bfe7ce
new print properties
