+++ pyoints/misc.py
@@ -140,20 +140,18 @@ def print_rounded(values, decimals=2):
     Sets negative values close to zero to zero.
 
     """
-    if isinstance(values, Number):
-        values = np.round(values, decimals=decimals)
-        if np.isclose(values, 0):
-            values = values * 0
-    else:
-        values[np.isclose(values, 0)] = 0
 
-    max_val = 5
-    import re
-    max_str_len = len(str(int(max_val))) + decimals + 2
-    formatter = {
-        'int_kind': lambda x:'{0:d}'.format(x),
-        'float_kind': lambda x: ('{0. <%i}|' % decimals).format(x),
-    }
+    if isinstance(values, Number):
+        rounded = np.round(values, decimals=decimals)
+        if np.isclose(rounded, 0):
+            rounded = rounded * 0
+    else:
+        rounded = np.round(values, decimals=decimals)
+        rounded[np.isclose(rounded, 0)] = 0
+        
+    if isinstance(values, tuple):
+        rounded = tuple(rounded)
+        
     np.set_printoptions(
         linewidth=75,
         precision=decimals,
@@ -161,21 +159,7 @@ def print_rounded(values, decimals=2):
         threshold=20,
         sign=' ',
         floatmode='unique',
-        #formatter=formatter,
         legacy='1.13'
     )
-
-    s = np.array2string(
-        values,
-        max_line_width=75,
-        precision=decimals,
-        suppress_small=True,
-        separator=' ',
-        threshold=20,
-        sign=' ',
-        floatmode='unique',
-        #formatter=formatter,
-        legacy='1.13',
-    )
-    print(values)
+    print(rounded)
 
