### 2019-02-27 5216b87
new print properties
