+++ pyoints/transformation.py
@@ -41,7 +41,6 @@ from .assertion import (
     isnumeric,
 )
 from .misc import print_rounded
-print_rounded(1)
 
 
 def transform(coords, T, inverse=False, precise=False):
@@ -72,16 +71,16 @@ def transform(coords, T, inverse=False, precise=False):
     >>> T = [(1, 0, 5), (0, 1, 3), (0, 0, 1)]
 
     >>> tcoords = transform(coords, T)
-    >>> print(tcoords)
-    [[5. 3.]
-     [6. 5.]
-     [7. 6.]
-     [5. 6.]]
-    >>> print(transform(tcoords, T, inverse=True))
-    [[0. 0.]
-     [1. 2.]
-     [2. 3.]
-     [0. 3.]]
+    >>> print_rounded(tcoords)
+    [[ 5.  3.]
+     [ 6.  5.]
+     [ 7.  6.]
+     [ 5.  6.]]
+    >>> print_rounded(transform(tcoords, T, inverse=True))
+    [[ 0.  0.]
+     [ 1.  2.]
+     [ 2.  3.]
+     [ 0.  3.]]
 
     """
     T = ensure_tmatrix(T)
@@ -241,11 +240,11 @@ def i_matrix(dim):
     Examples
     --------
 
-    >>> print(i_matrix(3))
-    [[1. 0. 0. 0.]
-     [0. 1. 0. 0.]
-     [0. 0. 1. 0.]
-     [0. 0. 0. 1.]]
+    >>> print_rounded(i_matrix(3))
+    [[ 1.  0.  0.  0.]
+     [ 0.  1.  0.  0.]
+     [ 0.  0.  1.  0.]
+     [ 0.  0.  0.  1.]]
 
     """
     if not (isinstance(dim, int) and dim >= 2):
@@ -275,11 +274,11 @@ def t_matrix(t):
     Examples
     --------
 
-    >>> print(t_matrix([1, 2, 3]))
-    [[1. 0. 0. 1.]
-     [0. 1. 0. 2.]
-     [0. 0. 1. 3.]
-     [0. 0. 0. 1.]]
+    >>> print_rounded(t_matrix([1, 2, 3]))
+    [[ 1.  0.  0.  1.]
+     [ 0.  1.  0.  2.]
+     [ 0.  0.  1.  3.]
+     [ 0.  0.  0.  1.]]
 
     """
     t = ensure_numvector(t, min_length=2)
@@ -309,7 +308,7 @@ def s_matrix(s):
     Examples
     --------
 
-    >>> print(s_matrix([0.5, 1, -3]))
+    >>> print_rounded(s_matrix([0.5, 1, -3]))
     [[ 0.5  0.   0.   0. ]
      [ 0.   1.   0.   0. ]
      [ 0.   0.  -3.   0. ]
@@ -358,7 +357,7 @@ def r_matrix(a, order='xyz'):
     Two dimensional case.
 
     >>> R = r_matrix(np.pi/4)
-    >>> print(R, 3)
+    >>> print_rounded(R, 3)
     [[ 0.707 -0.707  0.   ]
      [ 0.707  0.707  0.   ]
      [ 0.     0.     1.   ]]
@@ -366,7 +365,7 @@ def r_matrix(a, order='xyz'):
     Three dimensional case.
 
     >>> R = r_matrix([np.pi/2, 0, np.pi/4])
-    >>> print(R, 3)
+    >>> print_rounded(R, 3)
     [[ 0.707  0.     0.707  0.   ]
      [ 0.707  0.    -0.707  0.   ]
      [ 0.     1.     0.     0.   ]
@@ -447,17 +446,17 @@ def q_matrix(q):
[…]
