### 2019-02-27 f8653be
new print properties
