+++ pyoints/projection.py
@@ -179,18 +179,18 @@ class GeoTransform:
 
     >>> tCoords = geoTransfrom(coords)
     >>> print_rounded(tCoords, 3)
-    [[2560446.801 5537522.386]
-     [2560808.009 5537303.984]
-     [2564724.211 5539797.116]
-     [2566007.32  5552049.646]]
+    [[ 2560446.801  5537522.386]
+     [ 2560808.009  5537303.984]
+     [ 2564724.211  5539797.116]
+     [ 2566007.32   5552049.646]]
 
     Reverse transformation.
 
     >>> print_rounded(geoTransfrom(tCoords, reverse=True), 3)
-    [[ 6.842 49.971]
-     [ 6.847 49.969]
-     [ 6.902 49.991]
-     [ 6.922 50.101]]
+    [[  6.842  49.971]
+     [  6.847  49.969]
+     [  6.902  49.991]
+     [  6.922  50.101]]
 
     """
 
