### 2019-02-27 2bf8b39
fix of manual conda installation instructions
