### 2019-02-27 2507bba
new print properties
