### 2019-02-27 6ed81d1
new print properties
