+++ pyoints/grid/grid.py
@@ -75,32 +75,32 @@ class Grid(GeoRecords):
     >>> raster = Grid(proj, data, T)
     >>> print(raster.dtype.descr)
     [('values', '<i8'), ('coords', '<f8', (2,))]
-    >>> print(raster.shape)
+    >>> print_rounded(raster.shape)
     (4, 3)
-    >>> print(raster.dim)
+    >>> print_rounded(raster.dim)
     2
-    >>> print(raster.t.origin)
-    [10. 20.]
+    >>> print_rounded(raster.t.origin)
+    [ 10.  20.]
 
     Get cell data of the raster.
 
-    >>> print(raster.coords)
-    [[[10.25 20.2 ]
-      [10.75 20.2 ]
-      [11.25 20.2 ]]
+    >>> print_rounded(raster.coords)
+    [[[ 10.25  20.2 ]
+      [ 10.75  20.2 ]
+      [ 11.25  20.2 ]]
     <BLANKLINE>
-     [[10.25 20.6 ]
-      [10.75 20.6 ]
-      [11.25 20.6 ]]
+     [[ 10.25  20.6 ]
+      [ 10.75  20.6 ]
+      [ 11.25  20.6 ]]
     <BLANKLINE>
-     [[10.25 21.  ]
-      [10.75 21.  ]
-      [11.25 21.  ]]
+     [[ 10.25  21.  ]
+      [ 10.75  21.  ]
+      [ 11.25  21.  ]]
     <BLANKLINE>
-     [[10.25 21.4 ]
-      [10.75 21.4 ]
-      [11.25 21.4 ]]]
-    >>> print(raster.values)
+     [[ 10.25  21.4 ]
+      [ 10.75  21.4 ]
+      [ 11.25  21.4 ]]]
+    >>> print_rounded(raster.values)
     [[ 0  1  2]
      [ 3  4  5]
      [ 6  7  8]
@@ -108,12 +108,12 @@ class Grid(GeoRecords):
 
     Convert coordinates to indices and reverse.
 
-    >>> print(raster.coords_to_keys(raster.t.origin))
+    >>> print_rounded(raster.coords_to_keys(raster.t.origin))
     [0 0]
-    >>> print(raster.keys_to_coords([-0.5, -0.5]))
-    [10. 20.]
+    >>> print_rounded(raster.keys_to_coords([-0.5, -0.5]))
+    [ 10.  20.]
 
-    >>> print(raster.coords_to_keys(raster.coords))
+    >>> print_rounded(raster.coords_to_keys(raster.coords))
     [[[0 0]
       [0 1]
       [0 2]]
@@ -129,31 +129,31 @@ class Grid(GeoRecords):
      [[3 0]
       [3 1]
       [3 2]]]
-    >>> print(raster.keys_to_coords(raster.keys))
-    [[[10.25 20.2 ]
-      [10.75 20.2 ]
-      [11.25 20.2 ]]
+    >>> print_rounded(raster.keys_to_coords(raster.keys))
+    [[[ 10.25  20.2 ]
+      [ 10.75  20.2 ]
+      [ 11.25  20.2 ]]
     <BLANKLINE>
-     [[10.25 20.6 ]
-      [10.75 20.6 ]
-      [11.25 20.6 ]]
+     [[ 10.25  20.6 ]
+      [ 10.75  20.6 ]
+      [ 11.25  20.6 ]]
     <BLANKLINE>
-     [[10.25 21.  ]
-      [10.75 21.  ]
-      [11.25 21.  ]]
+     [[ 10.25  21.  ]
+      [ 10.75  21.  ]
+      [ 11.25  21.  ]]
     <BLANKLINE>
[…]
