### 2019-02-27 88f720b
new print properties
