### 2019-02-27 cfdb134
new print properties
