### 2019-02-27 laempy/pyoints#8 @laempy
Thanks for these hints, but both doi's seem not to match.

https://doi.org/10.1109/mra.2009.933612 points to a textbook with title "Learning OpenCV". So, it isn't the intended reference. I have the bibtex entry from https://github.com/opencv/opencv/wiki/CiteOpenCV.
https://doi.org/10.1109/cvpr.2018.00409 points to a paper called "Tangent Convolutions for Dense Prediction in 3D". It just cites Open3D.
