### 2019-02-28 8938a8c
installation instructions updated
