### 2019-02-28 e7d3c9c
readme adapted to zendo citation
