+++ README.md
@@ -169,17 +169,9 @@ License v3 or later (GPLv3+). See [LICENSE](LICENSE) for details.
 Please cite *Pyoints* as follows:
 
 S. Lamprecht (2019):
-"Pyoints: A Python package for point cloud, voxel and raster processing". [![DOI](https://zenodo.org/badge/144562205.svg)](https://zenodo.org/badge/latestdoi/144562205)
+"Pyoints: A Python package for point cloud, voxel and raster processing".
+[![DOI](https://zenodo.org/badge/144562205.svg)](https://zenodo.org/badge/latestdoi/144562205)
 
-```
-@Misc{lamprecht_2019,
-  author = {Sebastian Lamprecht},
-  organization = {Trier University},
-  title = {{Pyoints}: A Python package for point cloud, voxel and raster processing.},
-  year = {2019},
-  doi = {10.5281/zenodo.2557575}
-}
-```
 
 
 ## Contributing
