### 2019-03-01 e4b7274
Handling of JSON files enabled
