### 2019-03-01 fc2af49
some instructions on conda
