### 2019-03-01 c4569b8
simplification of meta.yaml
