+++ pyoints/georecords.py
@@ -21,8 +21,7 @@
 
 import warnings
 import numpy as np
-
-from .coords import Coords
+from datetime import datetime
 
 from . import (
     transformation,
@@ -30,6 +29,7 @@ from . import (
     assertion,
     nptools,
 )
+from .coords import Coords
 
 from .misc import print_rounded
 
@@ -66,6 +66,8 @@ class GeoRecords(np.recarray, object):
         of a point cloud or number of cells of a raster.
     keys : np.ndarray(int, shape=self.shape)
         Keys or indices of the data structure.
+    date : datetime
+        Date of capture.
 
     Examples
     --------
@@ -120,13 +122,15 @@ class GeoRecords(np.recarray, object):
 
     """
 
-    def __init__(self, proj, rec, T=None):
-        self.proj = proj    # validated by setter
+    def __init__(self, proj, rec, T=None, date=None):
         if T is None:
             T = transformation.t_matrix(self.extent().min_corner)
-        self.t = T    # validated by setter
+        # validated by setter
+        self.proj = proj
+        self.t = T
+        self.date = date
 
-    def __new__(cls, proj, rec, T=None):
+    def __new__(cls, proj, rec, T=None, date=None):
         if isinstance(rec, dict):
             rec = nptools.recarray(rec)
         elif not isinstance(rec, np.recarray):
@@ -143,28 +147,26 @@ class GeoRecords(np.recarray, object):
     def dim(self):
         return self.dtype['coords'].shape[0]
 
-
     @property
     def coords(self):
         #return self['coords'].view(Coords)
         if not hasattr(self, '_coords'):
-        #    # copy required for garbadge collection
+            # copy required for garbadge collection
             self._coords = self['coords'].copy().view(Coords)
         return self._coords
-    
-        
+
     def _clear_cache(self):
         if hasattr(self, '_coords'):
             del self._coords
 
     def __setattr__(self, attr, value):
         np.recarray.__setattr__(self, attr, value)
-        if attr is 'coords':
+        if attr == 'coords':
             self._clear_cache()
 
     def __setitem__(self, key, value):
         np.recarray.__setitem__(self, key, value)
-        if key is 'coords':
+        if key == 'coords':
             self._clear_cache()
 
     def extent(self, *args, **kwargs):
@@ -178,6 +180,7 @@ class GeoRecords(np.recarray, object):
             return
         self._t = getattr(obj, '_t', None)
         self._proj = getattr(obj, '_proj', None)
+        self._date = getattr(obj, '__date', None)
 
     # def __array_wrap__(self, out_arr, context=None):
     #     return np.ndarray.__array_wrap__(self, out_arr, context)
@@ -210,6 +213,17 @@ class GeoRecords(np.recarray, object):
     @property
     def count(self):
         return np.product(self.shape)
+
+    @property
+    def date(self):
+        return self._date
+
+    @date.setter
+    def date(self, date):
+        if (date is not None) and (not isinstance(date, datetime)):
+            m = "'date' needs to be of type 'datetime', got %s" % type(date)
+            raise TypeError(m)
+        self._date = date
 
     def records(self):
         """Provides the flattened data records. Useful if structured data
@@ -397,7 +411,7 @@ class GeoRecords(np.recarray, object):
         """
 
         records = nptools.add_fields(self, dtypes, data=data)
-        return self.__class__(self.proj, records, T=self.t)
+        return self.__class__(self.proj, records, T=self.t, date=self.date)
 
     def merge(self, rec):
         """Merges a record array with the georecords.
@@ -418,7 +432,7 @@ class GeoRecords(np.recarray, object):
 
         """
         data = nptools.merge((self, rec))
[…]
