### 2019-03-04 42e081f
Installation instructions revised
