+++ README.md
@@ -81,15 +81,35 @@ Conda installation is recommended. To install *Pyoints* with
 conda install -c laempy pyoints
 ```
 
-Currently Linux (64 bit) and Windows (64 bit) have been packaged.
+Currently only Linux (64 bit) has been packaged with Python 3.5, 3.6 and 3.7.
+
+
+### Installation from source
+
+We recommend to use a virtual environment (like Conda) to install *Pyoints*.
+
+Installation of the external dependencies:
+```
+conda install gdal pyproj numpy>=1.15 rtree
+```
+
+Cloning of the repository:
+```
+git clone https://github.com/laempy/pyoints.git
+cd pyoints
+```
+
+Installation of the python dependencies via *pip*.
+```
+pip install -e .
+```
 
 
 ### Conda Build
 
 If your system is not supported yet, you might build and install it using the
 [conda reciepe](conda/meta.yaml) specifying your desired Python version.
-Currently Linux (64 bit) and Windows (64 bit) have been tested successfully
-with Python 3.5, 3.6 and 3.7.
+
 
 #### Build
 
@@ -120,30 +140,6 @@ Test the package by running the provided doctests:
 ```
 python /path/to/pyoints/tests/test_pyoints.py
 ```
-
-
-### Installation from source
-
-We recommend to use a virtual environment (like Conda) to install *Pyoints*.
-
-Installation of the external dependencies:
-```
-conda install gdal pyproj numpy rtree
-```
-
-Cloning of the repository:
-```
-git clone https://github.com/laempy/pyoints.git
-cd pyoints
-```
-
-Installation of the python packages via *pip*.
-```
-pip install pygdal==$(gdal-config --version).* -e .
-```
-Unfortunately, the gdal version is not detected automatically by *pygdal*.
-Thus, we have to specify the gdal version expicilely with
-`gdal-config --version`.
 
 
 ## Documentation
