+++ README.md
@@ -83,6 +83,13 @@ conda install -c laempy pyoints
 
 Currently only Linux (64 bit) has been packaged with Python 3.5, 3.6 and 3.7.
 
+To enable full LAS support (spatial reference system definition), you need to 
+install *liblas* and its python bindings manually.
+
+```
+pip install liblas
+```
+
 
 ### Installation from source
 
@@ -131,13 +138,6 @@ Finally install the package:
 ```
 conda activate pyoints_env
 conda install pyoints
-```
-
-To enable full LAS support (spatial reference system definition), you need to 
-install *liblas* and its python bindings manually.
-
-```
-pip install liblas
 ```
 
 
