### 2019-03-04 c04dd4c
Docs for 0.2.0rc1
