### 2019-03-04 62d3518
system dependent datatype represenation removed from docstrings
