+++ pyoints/examples/file_examples/las_example.py
@@ -34,8 +34,8 @@ Create GeoRecords from scratch.
 ...                     center=[332592.88, 5513244.80, 120], epsg=25832)
 >>> print(geoRecords.shape)
 (1000,)
->>> print(sorted(geoRecords.dtype.descr))
-[('classification', '<u8'), ('coords', '<f8', (3,)), ('intensity', '<u8'), ('keypoint', '|b1'), ('synthetic', '|b1'), ('values', '<f8'), ('withheld', '|b1')]
+>>> print(sorted(geoRecords.dtype.names))
+['classification', 'coords', 'intensity', 'keypoint', 'synthetic', 'values', 'withheld']
 
 Set the capturing date.
 
