### 2019-03-04 #990 @laempy
@lheagy: I have taken a look at the DOIs (see laempy/pyoints#8) and created a new release **0.2.0rc2** with DOI 10.5281/zenodo.2582977 (or 10.5281/zenodo.2557574 for citing all versions). I have noticed, that the PDF file created by whedon is based on an old version of the package. So, it needs to be regenerated for proof reading.
After several problems with the conda recipe I had to drop Windows support. So I have created conda packages for linux only. But, the manual conda installation should still work (tested with win7, 64-bit).
