### 2019-03-06 a1cbb97
import of about parameters revised
