### 2019-03-06 a9e62e4
copyright sring revised
