+++ docs/_sources/pyoints.rst.txt
@@ -14,6 +14,14 @@ Subpackages
 Submodules
 ----------
 
+pyoints.about module
+--------------------
+
+.. automodule:: pyoints.about
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 pyoints.assertion module
 ------------------------
 
