### 2019-03-06 #990 @laempy
Hi @lheagy and @kbevers, I have found some remaining issues. So, I created the release **0.2.0rc3** with DOI 10.5281/zenodo.2585430. Could you please take another look? Thanks a lot.
