### 2019-03-07 de5a108
optimized serialization
