### 2019-03-17 #990 @lheagy
👋 Hi @laempy, just checking in - would you mind updating the zenodo archive record so that the title and author list are the same as on the paper?
