+++ paper/paper.bib
@@ -7,9 +7,9 @@
 
 @online{Pyoints_docs,
   author = {Sebastian Lamprecht},
-  title = {Pyoints},
+  title = {Pyoints Documentation},
   year = {2019},
-  url = {https://laempy.github.io/pyoints/pyoints.html},
+  url = {https://laempy.github.io/pyoints/},
   urldate = {2019-02-05}
 }
 
