+++ pyoints/storage/RasterHandler.py
@@ -68,8 +68,9 @@ class RasterReader(GeoFile):
         if proj is None:
             wkt = gdalRaster.GetProjection()
             if not wkt == '':
-                self.proj = projection.Proj.from_proj4(
-                    osr.SpatialReference(wkt=wkt).ExportToProj4())
+                proj = projection.Proj.from_wkt(wkt)
+            else:
+                warnings.warn("no projection found")
         self.proj = proj
 
         self.t = transformation.matrix_from_gdal(gdalRaster.GetGeoTransform())
