### 2019-03-29 #990 @danielskatz
👋 @laempy - I've suggested a bunch of changes in the paper text in https://github.com/laempy/pyoints/pull/12 - please make whichever of them you think are useful.  Most are to make the English more idiomatic, some are spelling and grammar fixes as well.
