### 2019-04-01 79a4955
autopep8 and licence note updated
