+++ pyoints/indexkd.py
@@ -34,6 +34,7 @@ from . import (
 
 from .misc import print_rounded
 
+
 class IndexKD(object):
     """Wrapper class for several spatial indices to speed up spatial queries
     and ease the usage.
