+++ pyoints/storage/dtype_converters.py
@@ -1,3 +1,21 @@
+# BEGIN OF LICENSE NOTE
+# This file is part of Pyoints.
+# Copyright (c) 2018, Sebastian Lamprecht, Trier University,
+# lamprecht@uni-trier.de
+#
+# Pyoints is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# Pyoints is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Pyoints. If not, see <https://www.gnu.org/licenses/>.
+# END OF LICENSE NOTE
 """Helper functions for data type conersion.
 """
 
