+++ docs/_downloads/7dfed88757b76770b18d6dc7cb5e533e/stem_detection.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+
 # coding: utf-8
 
 # # Detection of tree stems with Pyoints
