+++ docs/_downloads/d9c564c5802e6e3aee0ea2d72df062e9/icp.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+
 # coding: utf-8
 
 # # ICP for point cloud alignment
