+++ docs/_downloads/61a40433bbbf8ad0e26addb57a9fa5fd/getting_started.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+
 # coding: utf-8
 
 # # Getting started
