### 2019-04-01 #990 @laempy
Thank you @lheagy for accepting my paper. @danielskatz - Thank you for the suggestions in laempy/pyoints#12. I am going to add your modifications and create the final release. I will post the corresponding zenodo DOI here as soon as possible.
